+++ modules/Sources/Features/ExportTransactionHistory/ExportTransactionHistoryView.swift
@@ -50,7 +50,7 @@ public struct ExportTransactionHistoryView: View {
                 }
             }
             .zashiBack()
-            .walletStatusPanel()
+            //..walletstatusPanel()
 
             shareLogsView()
         }