+++ modules/Sources/UIComponents/Text/WalletStatusPanel.swift
@@ -81,7 +81,7 @@ private struct WalletStatusPanel: View {
             Text("Hello, World")
         }
         .padding(.vertical, 1)
-        .walletStatusPanel()
+        //..walletstatusPanel()
         .navigationBarTitleDisplayMode(.inline)
         .navigationBarItems(
             trailing: Text("M")